requirements.txt
selenium
undetected-chromedriver
pymongo
python-dotenv
fake_useragent
requests
redis
dnspython